api de video

+++ infrastructure/src/main/java/com/fullcycle/admin/catalogo/infrastructure/services/impl/GCStorageService.java
@@ -6,7 +6,6 @@ import com.google.cloud.storage.BlobId;
 import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.StreamSupport;
@@ -16,35 +15,36 @@ public class GCStorageService implements StorageService {
     private final String bucket;
     private final Storage storage;
 
-    public GCStorageService(String bucket, Storage storage) {
+    public GCStorageService(final String bucket, final Storage storage) {
         this.bucket = bucket;
         this.storage = storage;
     }
 
     @Override
-    public void store(final String name, final Resource resource) {
-        final var blobInfo = BlobInfo.newBuilder(this.bucket, name)
+    public void store(final String id, final Resource resource) {
+        final var info = BlobInfo.newBuilder(this.bucket, id)
                 .setContentType(resource.contentType())
                 .setCrc32cFromHexString(resource.checksum())
                 .build();
 
-        this.storage.create(blobInfo, resource.content());
+        this.storage.create(info, resource.content());
     }
 
     @Override
-    public Optional<Resource> get(final String name) {
-        return Optional.ofNullable(this.storage.get(BlobId.of(bucket, name)))
+    public Optional<Resource> get(final String id) {
+        return Optional.ofNullable(this.storage.get(this.bucket, id))
                 .map(blob -> Resource.with(
-                        blob.getCrc32cToHexString(),
                         blob.getContent(),
+                        blob.getCrc32cToHexString(),
                         blob.getContentType(),
-                        name
+                        blob.getName()
                 ));
     }
 
     @Override
     public List<String> list(final String prefix) {
         final var blobs = this.storage.list(bucket, Storage.BlobListOption.prefix(prefix));
+
         return StreamSupport.stream(blobs.iterateAll().spliterator(), false)
                 .map(BlobInfo::getBlobId)
                 .map(BlobId::getName)
@@ -52,9 +52,12 @@ public class GCStorageService implements StorageService {
     }
 
     @Override
-    public void deleteAll(final Collection<String> names) {
-        final var blobs = names.stream()
-                .map(name -> BlobId.of(bucket, name))
+    public void deleteAll(final List<String> ids) {
+        if (ids == null || ids.isEmpty())
+            return;
+
+        final var blobs = ids.stream()
+                .map(id -> BlobId.of(this.bucket, id))
                 .toList();
 
         this.storage.delete(blobs);
